src/processing/requirements.txt
faust-streaming==0.10.14
aiokafka==0.8.1
redis==5.0.1
asyncpg==0.29.0
python-dotenv==1.0.0
numpy==1.26.2
pandas==2.1.3
